Update JWT service and user CRUD operation project structure modify (router)

+++ api/model/role.go
@@ -4,6 +4,6 @@ import "gorm.io/gorm"
 
 type Roles struct {
 	gorm.Model
-	ID   uint64 `gorm:"primaryKey"`
+	ID   uint8  `gorm:"primaryKey"`
 	Name string `gorm:"type:char(10);not null"`
 }
